requirements.txt
Flask
paho-mqtt
sqlalchemy
psycopg2